# Divide FLOPs by two due to causal masks