# Description

As pointed out by [this issue](https://github.com/AI-Hypercomputer/maxtext/issues/1972), the current FLOPs calculation does not consider the impact from causal masks. This PR fixes this issue.


FIXES: b/431892390

# Tests

| Model             | Max Target Length | TFLOPs Type      | Before       | After        |
| :---------------- | :---------------- | :--------------- | :----------- | :----------- |
| **llama2-7b** | 8192              | **Total** | **1721.22** | **1510.11** |
|                   |                   | Learnable Weight | 1299.00      | 1299.00      |
|                   |                   | Attention        | 422.21       | 211.11       |
|                   | 129024            | **Total** | **125194.40**| **72826.86** |
|                   |                   | Learnable Weight | 20459.32     | 20459.32     |
|                   |                   | Attention        | 104735.08    | 52367.54     |
| **deepseek2-16b** | 8192              | **Total** | **704.60** | **593.27** |
|                   |                   | Learnable Weight | 481.95       | 481.95       |
|                   |                   | Attention        | 222.65       | 111.33       |
|                   | 129024            | **Total** | **62822.05** | **35206.36** |
|                   |                   | Learnable Weight | 7590.66      | 7590.66      |
|                   |                   | Attention        | 55231.39     | 27615.69     |
| **llama3-8b** | 8192              | **Total** | **1897.69** | **1686.58** |
|                   |                   | Learnable Weight | 1475.48      | 1475.48      |
|                   |                   | Attention        | 422.21       | 211.11       |
|                   | 129024            | **Total** | **127973.82**| **75606.28** |
|                   |                   | Learnable Weight | 23238.75     | 23238.75     |
|                   |                   | Attention        | 104735.08    | 52367.54     |
| **llama4-17b-16e**| 8192              | **Total** | **3964.49** | **3568.67** |
|                   |                   | Learnable Weight | 3172.84      | 3172.84      |
|                   |                   | Attention        | 791.65       | 395.82       |
|                   | 129024            | **Total** | **246350.55**| **148161.41**|
|                   |                   | Learnable Weight | 49972.27     | 49972.27     |
|                   |                   | Attention        | 196378.27    | 98189.14     |

# Checklist

Before submitting this PR, please make sure (put X in square brackets):
- [x] I have performed a self-review of my code.
- [x] I have necessary comments in my code, particularly in hard-to-understand areas.
- [x] I have run end-to-end tests tests and provided workload links above if applicable.
- [x] I have made or will make corresponding changes to the doc if needed.